Add Jira ticket link support
This does nothing fancy..  if it sees something that looks like it might
be a Jira ticket number (e.g., 'QPPCP-3050') in a message, Flexibot will
respond with a link to that ticket in Jira.  It literally prepends a
base string ('https://jira.cms.gov/browse/') with the ticket number.

Oh, and there's another Easter Egg in there (which is obvious from the
git diff).