Only return links to tickets numbers, not URLs
Previously, if the user would provide one or more ticket numbers,
Flexibot would blindly create links to those tickets.  Cool.

However, if the user provided the URL to the ticket, Flexibot would
recreate the URL and respond with it... which wasn't helpful.  Also, if
the ticket number was already included in the response (from Flexibot),
Flexibot would say it again.. which, again, wasn't helpful.

So, now if the full URL to the ticket exists in either the message (from
the user) or the response (from Flexibot), the URL will not be included
in the response.

Also, the method for creating the response was cleaned-up to be a little
more legible and/or useful for extension.