Separate, test, and fix off-hours check
First, the logic for off-hours checking was placed in a separate file
which is imported at runtime.

This allowed me to run unit tests on the logic for each test.

This showed where is_offhours was potentially failing due to a Date
object not being passed consistently.

This also highlighted where local and UTC times were being used
incorrectly (always use getUTCxxx() so we know that the hour is in UTC
and we can correctly and consistently use timezone offsets.

While I was at it, I didn't like having the logic for whether or not a
date was a weekday or weekend in the is_offhours() function; plus,
I couldn't check that logic separately.  So, a new function for testing
weekend vs weekday was created.

Similarly, I didn't like having the logic for the time of day checks
in is_offhours() and it couldn't be checked separately, so two new
functions were added for testing is_before_hours() and is_after_hours().

While writing is_after_hours(), I noticed that the check was for when
the hour was >= 5 being off-hours.  This is incorrect as the hours are
in 24 hour format; the correct logic was hour >= 17.