Refactor jira_matcher and add documentation
The documentation is self-explanatory.

The refactoring was to allow us to specify "class variables" that can
be used outside of the jira_matcher (i.e., the jira_regex that is used
by Flexibot to determine if a string matches or not).  In doing so, the
syntax was cleaned up a bit and the functions are now actual functions,
not lambdae pointed to by exported properties.